requirements.txt
absl-py==2.1.0
attrs==24.2.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cmudict==1.0.31
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
# decord==0.6.0
docker-pycreds==0.4.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
# NVIDIA CUDA packages (only for Linux/Windows with NVIDIA GPU)
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.6.20
# nvidia-nvtx-cu12==12.1.105
openai-whisper
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
pronouncing==0.2.0
protobuf==4.25.4
psutil==6.0.0
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.2
scipy==1.13.1
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
# triton  # Only for NVIDIA GPUs on Linux
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.17.9
zipp==3.20.1
